backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "smart-agent-backend"
version = "0.1.0"
description = "Smart Agent Platform backend powered by FastAPI and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "sumoer", email = "sumoqing2@gmail.com" }]
dependencies = [
    "fastapi>=0.115.2,<1.0",
    "starlette>=0.40.0,<1.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.18",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",
    "langgraph-checkpoint-postgres==2.0.25",
    "mysql-connector-python>=8.2.0",
    "pymilvus>=2.3.4",
    "asyncpg>=0.29.0",
    "neo4j>=5.20.0",
    "pypdf>=3.0.0",
    "langchain-text-splitters>=0.0.1",
    "transformers>=4.21.0",
    "torch>=1.12.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.2.1",
    "python-dateutil>=2.8.2",
    "tenacity>=9.0.0",
    "httpx>=0.25.2",
    "aiofiles>=23.2.1",
    "openai>=1.30.0",
    "hanlp>=2.1.0",
    "pyahocorasick>=2.0.0",
    "openai-whisper>=20231117",
    "PyJWT>=2.8.0",
    "psycopg-binary>=3.2.10",
    "psycopg-pool>=3.2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "yapf==0.40.2",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
